requirements.txt
# DocuMind - AI Document Intelligence Agent
# Core dependencies

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.0
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# OCR
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3

# NLP and AI
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.20
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# Vector Store and Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4
numpy==1.26.3

# Data Processing
pandas==2.1.4
tabula-py==2.9.0
camelot-py[cv]==0.11.0

# Date and Entity Extraction
dateparser==1.2.0
spacy==3.7.2
python-dateutil==2.8.2

# Evaluation
rouge-score==0.1.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0
tqdm==4.66.1

# Memory and Storage
pickle5==0.0.12
json5==0.9.14

# Logging and Observability
loguru==0.7.2

# Web Framework (for API if needed)
flask==3.0.0
flask-cors==4.0.0

